Refactor games services
