+++ src/controllers/users_controllers/users_controller.js
@@ -1,7 +1,8 @@
 const express = require('express');
-const {getUsersByUsername} = require("../../services/users_services/users_getter_service");
-const {asyncWrapper} = require("../../utils/async_wrapper");
-const {selfUserRouter} = require("./self_user_controller");
+const {getUsersByUsername} =
+    require('../../services/users_services/users_getter_service');
+const {asyncWrapper} = require('../../utils/async_wrapper');
+const {selfUserRouter} = require('./self_user_controller');
 const {authMiddleware} = require('../../middlewares/auth_middleware');
 
 // eslint-disable-next-line new-cap
