+++ src/services/users_services/users_editing_service.js
@@ -1,7 +1,7 @@
-const {encryptPassword} = require("../../utils/password_processing");
-const {User} = require("../../models/user_model");
-const {deleteUserGamesRecord} = require("./user_games_service");
-const {deleteUserFriendsRecord} = require("./friends_service");
+const {encryptPassword} = require('../../utils/password_processing');
+const {User} = require('../../models/user_model');
+const {deleteUserGamesRecord} = require('../games_services/user_games_service');
+const {deleteUserFriendsRecord} = require('./friends_service');
 
 const deleteUserById = async (id) => {
     await User.deleteOne({_id: id});
